Struct massa_consensus::BootstrapableGraph

pub struct BootstrapableGraph {
    pub active_blocks: BlockHashMap<ExportActiveBlock>,
    pub best_parents: Vec<(BlockId, u64)>,
    pub latest_final_blocks_periods: Vec<(BlockId, u64)>,
    pub gi_head: BlockHashMap<BlockHashSet>,
    pub max_cliques: Vec<Clique>,
    pub ledger: LedgerSubset,
}

Fields
active_blocks: BlockHashMap<ExportActiveBlock>

Map of active blocks, were blocks are in their exported version.

best_parents: Vec<(BlockId, u64)>

Best parents hashe in each thread.

latest_final_blocks_periods: Vec<(BlockId, u64)>

Latest final period and block hash in each thread.

gi_head: BlockHashMap<BlockHashSet>

Head of the incompatibility graph.

max_cliques: Vec<Clique>

List of maximal cliques of compatible blocks.

ledger: LedgerSubset

Ledger at last final blocks