Struct massa_consensus::ConsensusCommandSender

pub struct ConsensusCommandSender(pub Sender<ConsensusCommand>);

Tuple Fields
0: Sender<ConsensusCommand>
Implementations

impl ConsensusCommandSender

pub async fn get_block_graph_status(
&self,
slot_start: Option<Slot>,
slot_end: Option<Slot>
) -> Result<BlockGraphExport, ConsensusError>

Gets all the aviable information on the block graph returning a Blockgraphexport.

Arguments

slot_start: optional slot start for slot-based filtering (included).

slot_end: optional slot end for slot-based filtering (excluded).

pub async fn get_cliques(&self) -> Result<Vec<Clique>, ConsensusError>

Gets all cliques.

pub async fn get_block_status(
&self,
block_id: BlockId
) -> Result<Option<ExportBlockStatus>, ConsensusError>

Gets the whole block and its status corresponding to given hash.

Arguments

hash: hash corresponding to the block we want.

pub async fn get_active_block(
&self,
block_id: BlockId
) -> Result<Option<Block>, ConsensusError>

Gets the whole block corresponding to given hash.

Arguments

hash: hash corresponding to the block we want.

pub async fn get_selection_draws(
&self,
start: Slot,
end: Slot
) -> Result<Vec<(Slot, (Address, Vec<Address>))>, ConsensusError>

Gets (slot, public_key) were the staker with public_key was selected for slot, between start_slot and end_slot.

Arguments

start_slot: begining of the considered interval.

end_slot: end of the considered interval.

pub async fn get_bootstrap_state(
&self
) -> Result<(ExportProofOfStake, BootstrapableGraph), ConsensusError>

pub async fn get_block_ids_by_creator(
&self,
address: Address
) -> Result<BlockHashMap<Status>, ConsensusError>

pub async fn get_operations(
&self,
operation_ids: OperationHashSet
) -> Result<OperationHashMap<OperationSearchResult>, ConsensusError>

pub async fn get_addresses_info(
&self,
addresses: AddressHashSet
) -> Result<AddressHashMap<AddressState>, ConsensusError>

Gets the candidate and final ledger data of a list of addresses

pub async fn get_operations_involving_address(
&self,
address: Address
) -> Result<OperationHashMap<OperationSearchResult>, ConsensusError>

Returns hashmap: Operation id -> if it is final

pub async fn get_stats(&self) -> Result<ConsensusStats, ConsensusError>

pub async fn get_active_stakers(
&self
) -> Result<AddressHashMap<u64>, ConsensusError>

pub async fn register_staking_private_keys(
&self,
keys: Vec<PrivateKey>
) -> Result<(), ConsensusError>

pub async fn remove_staking_addresses(
&self,
addresses: AddressHashSet
) -> Result<(), ConsensusError>

pub async fn get_staking_addresses(
&self
) -> Result<AddressHashSet, ConsensusError>

pub async fn get_stakers_production_stats(
&self,
addrs: AddressHashSet
) -> Result<Vec<StakersCycleProductionStats>, ConsensusError>

pub async fn get_endorsements_by_address(
&self,
address: Address
) -> Result<EndorsementHashMap<Endorsement>, ConsensusError>

pub async fn get_endorsements_by_id(
&self,
endorsements: EndorsementHashSet
) -> Result<EndorsementHashMap<EndorsementInfo>, ConsensusError>